pages and components path rearraged
