+++ src/App.js
@@ -1,16 +1,16 @@
 import { Route, Routes } from 'react-router-dom';
 import './App.css';
-import AddAdmin from './Pages/AddAdmin/AddAdmin';
-import AddDonation from './Pages/AddDonation/AddDonation';
-import AllDonations from './Pages/AllDonations/AllDonations';
-import Dashboard from './Pages/Dashboard/Dashboard';
-import Donate from './Pages/Donate/Donate';
+import AddAdmin from './Pages/Dashboard/AddAdmin/AddAdmin';
+import AddDonation from './Pages/Dashboard/AddDonation/AddDonation';
+import AllDonations from './Pages/Dashboard/AllDonations/AllDonations';
+import DashboardMain from './Pages/Dashboard/DashboardMain/DashboardMain';
+import Donate from './Pages/Dashboard/Donate/Donate';
 import Home from './Pages/Homepage/Home/Home';
 import Login from './Pages/Login/Login';
 import NotFound from './Pages/NotFound/NotFound';
 import RegistrationForm from './Pages/RegistrationForm/RegistrationForm';
-import DonationList from './Pages/UserDashboard/DonationList/DonationList';
-import UserDashboard from './Pages/UserDashboard/UserDashboard';
+import DonationList from './Pages/Dashboard/UserDashboard/UserDonationList/UserDonationList';
+import UserDashboard from './Pages/Dashboard/UserDashboard/UserDashboard';
 import UserProfile from './Pages/UserProfile/UserProfile';
 
 function App() {
@@ -21,7 +21,7 @@ function App() {
         <Route path="/login" element={<Login />} />
         <Route path="/registration" element={<RegistrationForm />} />
         <Route path="/user-dashboard" element={<UserDashboard />} />
-        <Route path="/dashboard" element={<Dashboard />} />
+        <Route path="/dashboard" element={<DashboardMain />} />
         <Route path="/profile" element={<UserProfile />} />
         <Route path="/all-donations" element={<AllDonations />} />
         <Route path="/donate" element={<Donate />} />
