+++ src/Pages/Dashboard/AddAdmin/AddAdmin.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import Sidebar from '../../components/Sidebar/Sidebar';
+import Sidebar from '../../../components/Sidebar/Sidebar';
 
-const AddDonation = () => {
+const AddAdmin = () => {
   return (
     <div className="container-fluid">
       <div className="row g-md-2">
@@ -9,17 +9,9 @@ const AddDonation = () => {
           <Sidebar />
         </div>
         <div className="col-9 mt-2 mt-md-5 col-xl-10 px-sm-0 px-md-5 py-3 py-md-5" id="profile">
-          <h4 className="m-2">Add A Donation </h4>
+          <h4 className="m-2">Add an Admin: </h4>
           <p className="m-2">
-            <input type="text" name="donationName" id="addDonation" placeholder="Title" />
-          </p>
-
-          <p className="m-2">
-            <input type="file" name="uploadDonationPicture" />
-          </p>
-
-          <p className="m-2">
-            <input type="text" name="details" placeholder="Details" />
+            <input className="px-5" type="email" name="email" id="addAdmin" placeholder="Provide Admin's Email" />
           </p>
           <p className="m-2">
             <button className="btn btn-primary px-3" type="submit">
@@ -32,4 +24,4 @@ const AddDonation = () => {
   );
 };
 
-export default AddDonation;
+export default AddAdmin;
