+++ src/Pages/Dashboard/AllDonations/AllDonations.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import Sidebar from '../../components/Sidebar/Sidebar';
+import Sidebar from '../../../components/Sidebar/Sidebar';
 import AllDonationDetails from './AllDonationDetails';
 import './AllDonations.css';
