--- src/Pages/UserDashboard/UserDashboard.css
@@ -1,9 +1,0 @@
-.side-menu {
-  background-color: #f5f5f5;
-}
-
-@media screen and (max-width: 992px) {
-  .side-menu {
-    padding-top: 0px;
-  }
-}
